Move text to text_manager.
