+++ napari/layers/utils/text_manager.py
